Remove ratingView from the collection
Current behavior: Rating is bound to each photo in the collection every time you change the rating for any individual photo. (When you change the rating for any one photo, that new rating is bound to every photo in the collection.)
Expected behavior: Rating should only be bound to the photo that is displayed in the photo view when the rating is changed.